requirements.txt
# FinPilot AI — Python Dependencies
# Pinned for stability on HuggingFace Spaces (2 vCPU, 8GB RAM)

# ─── Core Framework ──────────────────────────────────────────────────────────
streamlit==1.32.0
fastapi==0.110.0
uvicorn[standard]==0.27.1

# ─── AI / LLM (MANDATORY for hackathon) ──────────────────────────────────────
openai==1.14.0

# ─── Data & Computation ──────────────────────────────────────────────────────
numpy==1.26.4
pandas==2.2.1
plotly==5.20.0

# ─── Database ────────────────────────────────────────────────────────────────
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
alembic==1.13.1

# ─── Auth ────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# ─── Notifications ───────────────────────────────────────────────────────────
twilio==8.13.0
sendgrid==6.11.0

# ─── Payments ────────────────────────────────────────────────────────────────
paypalrestsdk==1.13.3

# ─── Currency ────────────────────────────────────────────────────────────────
httpx==0.27.0

# ─── Config & Utils ──────────────────────────────────────────────────────────
pyyaml==6.0.1
python-dotenv==1.0.1

# ─── Reports ─────────────────────────────────────────────────────────────────
reportlab==4.1.0
Pillow==10.2.0

# ─── Process Manager (for Dockerfile dual-server) ────────────────────────────
supervisor==4.2.5